+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 - FEATURES.md aka general feature list
 - CHANGELOG.md Basic changelog from start of versioning
 - Icon to main nav exit option
+- 'Discard' button for CPU profile edit
 
 ### Changed
 - Current CPU frequency values are now periodically checked with active profile and rewritten if found not to be the same.
+- 'Save' button on CPU profile edit relocated
 
 ## [0.3.0] - 2019-09-11
 ### Added